Split TraceCollector between Common and Interpreters
This is required to leave TraceCollector::collect() inside main for
MemoryTrackier.
